Cache the dependencies of a node in the node structure
Dependencies are used when computing the set of variables, when
building recursive equations and when applying substitutions, so
caching them instead of traversing the type each time is a significant
win.  The cache is invalidated whenever a type is simplified, since
some nodes may have been erased.

This commit also improve slightly the simplification of a node. When
simplifying a node, we then simplify the direct-nodes of its
simplified version, not the original vdescr, since some direct-nodes
may have been removed during simplification.

Speed-up 11.9%.