Clean-up the tallying code
Simplify the code for the tallying:
  - use a single representation for constraints
  - constraints are maintained in a merged form
  - use sorted lists instead of OCaml set, since the
  tallying code spends most of its time iterating through
  constraint sets or converting back and forth between lists.

Unfortunately, some extra subtyping tests are added to maintain
the same exploration order of constraints, to ensure the same
printing order as the mainline code. It only impacts a few
examples (2 in sstt test suite and 1 in MLsem prototype.exe),
which are correct when checked manually (only the order of branches
in an intersection of arrow differs).
These (non costly) extra check can be removed later.

Speed-up:  5%.