General "order-changing" optimizations.
This commit regroups all the optimizations that impact (slightly)
the printing order of types with respect to main. It is added last,
but has the biggest impact:

- in Node, add fast-path cup/cap/diff so that if the argument is
physically equal to any/empty, the correct trivial result is returned.
This alone improves things by more than 20%.
- remove the subtyping checks in Tallying that where there only to keep
the order of exploration of constraint-sets identical to main.
- some expressions that allocate intermediary lists are refactored to
use the iterations of the collections.

Speed-up: 30%.