Rename the base type `Atoms` to `Enums`
The term "atom" is ambiguous since it can denote:
- the kind of constant constructors identified by their name:
`nil`, `true`, `false`, `A`, `red`, …
- the constituant of a component. For instance `node list` is
the atom of the tuple component, `(node * node)` is the atom
of the arrow component. This meaning comes from atoms in as units
of the DNF (a DNF is a union of intersection of positive and negative "atoms").

We rename the constant to `Enums` which is a well established name for such things.