Small changes in `Bdd`
Makes several small changes in `Bdd`:
- change the comparison function so that `Leaf` constructors are larger,
to make it consistent with the ordering of atoms (subtrees are larger than
their parent)
- implement a `node'` smart constructor which checks whether the atom
is smaller than its left and right subtrees. This is then used in `map_nodes`
and `simplify` which reconstruct nodes after transforming the atom arbitrarily.
If the atom is not well ordered, it reverts to using union and intersection to
properly rebuild the tree
- simplify the code of simplify. Instead of having a bdd with a single hole,
we use a zipper à la Huet to know how to reconstruct the full bdd from the missing node.
- implement a fast-path for conj and disj for empty and singleton lists.

All these do not impact performances but make the code more robust.