Add `any : t` and `empty : t` to the `Node` signature (#9)
* Add `any : t` and `empty : t` to the `Node` signature

Split the definition of `Node` in several parts so that
the `Node` module that is passed as argument to the `Vdescr.Make` functor
contains `any` and `empty` as types rather than functions.

The split allows a safe `PreNode` module to exists on the cycle of
mutually recursive modules, which allows the initialization of the
modules to proceed without error.

* Add internal documentation to node.ml

Excerpt from the GH discussion are added at the top of node.ml to
describe why it is doing thing that way.