Simplify Bdd.simplify
When simplifying the sub-BDDs, we do not need to reconstruct the whole context
but only the current branch (sub-BDDs of other branches can be considered empty).
This does not show any significative performance improvement because I think the
irrelevant context was cancelling away anyway when doing the Bdd diff, but still
it simplifies the implementation.