Implement `hash : t -> int` for all modules
Each component as well as `Descr` and `Vdescr` now support an efficient
hash function.
Partial hashes are stored in the data-structure.
This allows to create efficient `Hashtbl` of `VDescr`, and use them instead
of `Map.Make(VDescr)` (except for the subtyping where the persistent nature
of Map is important to restore a previous state).

This improves performances slightly (~10%), mainly because now comparisons
are much faster (since they take the hash into account).

Note: the only hashsing that is not incremental is the hashing of record atoms,
where the default polymorphic hash (`Hashtbl.hash`) is used. This does not
seem to hurt since in our examples, atoms are small, and the performances can
be recovered if necessary by using the same approach as for tags (sorted list
of bindings incrementally hashed).