Pretty-printer refactoring
The interface between the main pretty-printer and extensions is
simplified as follows. An extension only has to provide:

- an `any` type, which covers all values of that type (in practice this
will be).
- a conversion function `to_t` from a `Ty.t` to some arbitrary represenatation.
The conversion function is passed a function which can convert a type `Ty.t`
to its algebraic representation and must use it to call back into the main
conversion.
- a `map` function which can apply a function `f` to all the algebraic
representation descriptor stored within the arbitrary representation
- a `print` function

The way these features are provided is through an opaque `extension_node` type.
The latter is a GADT with an existential type variable, which is equivalent,
but syntactically a bit lighter than using first class modules. The extensions
only use the `builder` function and do not need manually build a value of this
type. There is also no need to expose a language of binders.

See the documentation for `Printer.builder`.

A feature that seems unused at the moment is the possibility for an extension to
define its own aliases. It was kept, but untested, since neither the test suite
nor MLSem exercise this code.

The pretty-printing of tests is unchanged. The compat for MLsem is minimal (renaming
of `*_params'` to `*_params`) and the pretty-printing of tests is almost unchanged
(ordering or sharing for two functions changes a bit, but the length of the printed
type is the same, and of course, the type are equivalent).