Pretty-printer refactoring (#15)
* Pretty-printer refactoring

The interface between the main pretty-printer and extensions is
simplified as follows. An extension only has to provide:

- a conversion function `to_t` from a `Ty.t` to some arbitrary representation.
The conversion function is passed a function which can convert a type `Ty.t`
to its algebraic representation and must use it to call back into the main
conversion.
- a `map` function which can apply a function `f` to all the algebraic
representation descriptor stored within the arbitrary representation
- a `print` function

The way these features are provided is through an opaque `extension_node` type.
The latter is a GADT with an existential type variable, which is equivalent,
but syntactically a bit lighter than using first class modules. The extensions
only use the `builder` function and do not need manually build a value of this
type. There is also no need to expose a language of binders.

See the documentation for `Printer.builder`.