Fix a potential bug in the backtracking table
The backtracking table maintains a copy of the recursive
call stack. When checking the emptyness of a (recursive)
type X, if X visited again (during its check) then the
copy of the stack is remembered. When returning to the intial
call to X, if it happens to be non-empty, all types on that
saved call stack must be invalidated: their status was
unknown (they were on the stack) while making the wrong
hypothesis that X was empty.

If X is visited several times during its check, we need
to remember all the stacks, not only the latest one.
This is fixed by this commit.