Generalize the Tag component
The Tag component is generalized to allow an arbitrary interpretation.
Properties such as monotonicity and preservation of cap and cup can be specified, and subtyping is computed accordingly.
When not specified, the default properties of a tag are "monotonic + preserves cap and cup", which amounts to an identity interpretation,
equivalent to the former definition of Tags.
The default handling of Tags in the Printer has been generalized: now it shows the DNF (like for arrows).

Breaking changes:
- TagComp.as_atom moved into Op.TagComp.as_atom. Now requires the tag to have an identity interpretation (Op.TagComp.is_identity).