Implement subtyping for polymorphic records
The `psi` function in `records.ml` implements precisely the algorithm in Castagna & Peyrot, OOPSLA2025, in place of the previous implementation that translated records to tuples. This is no longer possible since, in addition to row variables, polymorphism at the level of fields is handled with Field Type variables (which are distinct from Type Variables). Field Types are implemented using a compact representation composed of two BDDs.