Make tallying more modular to accomodate for different variable kinds
Several constraint modules can now be written, for different variable kinds (in particular those of interest are type, field and row variables). A functor creates constraint sets over variables of a uniform kind. Sets of constraints of several kinds are kept together in a record, and sets of sets of constraints are represented by a list of such records.