Record incremental hash (#21)
Changes the internal representation of maps of labels (used for record fields) to sorted lists. Access to individual fields is slower (linear instead of log) but this is seldom used, except to type field access. Operations that iterate through the whole set of fields are much faster now. This also have the good property that semantically equal sets are structurally equals, which means they can easily incrementally hashed, giving a `hash` in constant time which improves operations on `Bdd` of records.